Rename app entry module and restore confetti
