+++ presentation/app/__init__.py
@@ -1,3 +1,7 @@
 """Application package for HobbyPicker."""
-from .main import start_app
+
+# Re-export the GUI launcher so consumers can simply ``from presentation.app
+# import start_app`` without needing to know the internal module layout.
+from .gui import start_app
+
 __all__ = ["start_app"]
